Created a basic system to operate autonomous stoplight
This code is a work in progress, and is nowhere near done. In this system, the smart car knows exactly what it will do before arriving at the intersection. Once the intersection sensor sees the smart car, it texts the car a prompt. The car should stop when it receives the prompt, and reply to the intersection with a request( for example "Request: Go Straight"). The intersection should process the request, and if it deems it safe, (i.e. the other car is not approaching the intersection) it will approve the request and the car can continue. If it is unsafe, the intersection will deny the request, and the car will ask again in some time. This is just an idea, and I had some spare time over the weekend, so I started writing basic code. If anybody does not like the system explained above, understand that this is merely an idea I had and I wrote it down here.