Added an All Clear text command
After the vehicle is estopped, the signal now checks to see if the other car has passed. If it has passed, then the signal sends another text message to the smart car saying that it is all clear, and it can now resume it's operation. 

This also comes with its own flag